requirements.txt
Authlib==1.2.1
python-dotenv==1.0.0
Flask==2.3.3
requests==2.31.0
